# API/openapi.yaml
openapi: 3.1.0
info:
  title: Book Worm E-Store API
  version: 1.0.0
  description: |
    REST API for the Book Worm online bookstore platform.
    Supports guest browsing, member authentication, catalogue management,
    cart, checkout, orders, payments, shipping, reviews, and recommendations.
  contact:
    name: Book Worm Engineering
    email: api@bookworm.io
  license:
    name: Proprietary
servers:
  - url: https://api.bookworm.io/v1
    description: Production
  - url: https://api-staging.bookworm.io/v1
    description: Staging

# ─────────────────────────────────────────────────────────────────────────────
# Security Schemes
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT access token (15 min TTL). Obtain via POST /auth/login.

    GuestToken:
      type: apiKey
      in: header
      name: X-Guest-Token
      description: Stable guest session identifier for unauthenticated cart/checkout flows.

# ─────────────────────────────────────────────────────────────────────────────
# Reusable Parameters
# ─────────────────────────────────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      description: 1-based page number
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: "Sort field and direction, e.g. `publishedDate:desc`"
      schema:
        type: string
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    ISODateTime:
      type: string
      format: date-time
      example: "2026-07-21T14:30:00+05:30"
    ISODate:
      type: string
      format: date
      example: "2026-07-21"
    MoneyDTO:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
          description: Decimal string — avoids IEEE-754 precision loss
          example: "149.00"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO-4217 currency code
          example: INR
    PaginationDTO:
      type: object
      required: [page, size, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId, timestamp]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHENTICATED
                - FORBIDDEN
                - RESOURCE_NOT_FOUND
                - CONFLICT
                - OPTIMISTIC_LOCK_CONFLICT
                - RESOURCE_DELETED
                - BUSINESS_RULE_VIOLATION
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
            message:
              type: string
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
            correlationId:
              $ref: '#/components/schemas/UUID'
            timestamp:
              $ref: '#/components/schemas/ISODateTime'

    # ── Address ──────────────────────────────────────────────────────────────
...
    AuthorRefDTO:
      type: object
      required: [authorId, name]
      properties:
        authorId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
...
    AuthorSummaryDTO:
      type: object
      required: [authorId, name, isFollowed]
      properties:
        authorId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        photoUrl:
          type: string
          format: uri
          nullable: true
        isFollowed:
          type: boolean
    AuthorDetailResponse:
      type: object
      required: [authorId, name, isActive, followerCount, isFollowed, books]
      properties:
        authorId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        bio:
          type: string
          nullable: true
        photoUrl:
          type: string
          format: uri
          nullable: true
        isActive:
          type: boolean
        followerCount:
          type: integer
          minimum: 0
        isFollowed:
          type: boolean
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookSummaryDTO'
    CreateAuthorRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 300
        bio:
          type: string
          maxLength: 5000
        photoUrl:
          type: string
          format: uri
    UpdateAuthorRequest:
      allOf:
        - $ref: '#/components/schemas/CreateAuthorRequest'
        - type: object
          required: [version]
          properties:
            version:
              type: integer
              minimum: 1
    PublisherRefDTO:
      type: object
      required: [publisherId, name]
      properties:
        publisherId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
...
    CategoryRefDTO:
      type: object
      required: [categoryId, name, slug]
      properties:
        categoryId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        slug:
          type: string
...
    DeliveryEstimateDTO:
      type: object
      required: [estimatedDate, isGuaranteed]
      properties:
        estimatedDate:
          $ref: '#/components/schemas/ISODate'
        isGuaranteed:
          type: boolean
    FormatPriceDTO:
      type: object
      required: [bookFormatId, formatType, price]
      properties:
        bookFormatId:
          $ref: '#/components/schemas/UUID'
        formatType:
          type: string
          enum: [PAPERBACK, HARDCOVER, EBOOK]
        price:
          $ref: '#/components/schemas/MoneyDTO'
        isbn:
          type: string
          nullable: true
        pageCount:
          type: integer
          nullable: true
...
    BookSummaryDTO:
      type: object
      required: [bookId, title, authors, categories, formats, reviewCount]
      properties:
        bookId:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/AuthorRefDTO'
        publisher:
          $ref: '#/components/schemas/PublisherRefDTO'
          nullable: true
        coverImageUrl:
          type: string
          format: uri
          nullable: true
        language:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryRefDTO'
        formats:
          type: array
          items:
            $ref: '#/components/schemas/FormatPriceDTO'
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: integer
          minimum: 0
        deliveryEstimate:
          $ref: '#/components/schemas/DeliveryEstimateDTO'
          nullable: true
...
    FacetDTO:
      type: object
      required: [value, count]
      properties:
        value:
          type: string
        count:
          type: integer
          minimum: 0
    PriceRangeFacetDTO:
      type: object
      required: [label, min, max, count]
      properties:
        label:
          type: string
        min:
          type: number
        max:
          type: number
        count:
          type: integer
          minimum: 0
    SearchFacetsDTO:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/FacetDTO'
        formats:
          type: array
          items:
            $ref: '#/components/schemas/FacetDTO'
        languages:
          type: array
          items:
            $ref: '#/components/schemas/FacetDTO'
        priceRanges:
          type: array
          items:
            $ref: '#/components/schemas/PriceRangeFacetDTO'
    SearchResponse:
      type: object
      required: [query, data, pagination, facets]
      properties:
        query:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/BookSummaryDTO'
        pagination:
          $ref: '#/components/schemas/PaginationDTO'
        facets:
          $ref: '#/components/schemas/SearchFacetsDTO'

    # ── Wishlist ─────────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: Insufficient role/permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Conflict:
      description: Conflict — unique constraint or optimistic lock violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
security:
  - GuestToken: []

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /authors:
    get:
      tags: [Authors]
      summary: List authors
      operationId: listAuthors
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Partial name filter
        - name: isActive
          in: query
          schema:
            type: boolean
            default: true
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated author list
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuthorSummaryDTO'
                  pagination:
                    $ref: '#/components/schemas/PaginationDTO'
    post:
      tags: [Authors]
      summary: Create an author (admin)
      operationId: createAuthor
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthorRequest'
      responses:
        '201':
          description: Author created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /authors/{authorId}:
    parameters:
      - name: authorId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
    get:
      tags: [Authors]
      summary: Get author detail
      operationId: getAuthor
      responses:
        '200':
          description: Author detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Authors]
      summary: Update an author (admin)
      operationId: updateAuthor
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuthorRequest'
      responses:
        '200':
          description: Author updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Authors]
      summary: Delete an author (admin)
      operationId: deleteAuthor
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Author deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════════════════════════════════════════════════
  # Publishers
  # ══════════════════════════════════════════════════════════════════════════
...
  /search/books:
    get:
      tags: [Search]
      summary: Full-text book search with facets
      operationId: searchBooks
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search term (min 2 characters)
        - name: language
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [PAPERBACK, HARDCOVER, EBOOK]
        - name: minPrice
          in: query
          schema:
            type: number
            minimum: 0
        - name: maxPrice
          in: query
          schema:
            type: number
            minimum: 0
        - name: categorySlug
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, 'price:asc', 'price:desc', 'rating:desc']
            default: relevance
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Search results with facets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ══════════════════════════════════════════════════════════════════════════
  # Cart
  # ══════════════════════════════════════════════════════════════════════════
...